cthulhu: Refactor RequestCollection to be global
This is instead of having one per ClusterMonitor.  The
idea here is two accomplish two things:
* Expose RequestCollection to ServerMonitor
* Enable REST API to easily list all the requests

The alternative was to instantiate an additional RequestCollection
for servers.  This would have been awkward, because requests on
servers are also "in" zero or more clusters, so the API querying
for requests would have become complicated, touching multiple
RequestCollections.

A separate RequestCollection encapsulated in ServerMonitor would
also have lacked calls to its on_map method for notifications.  This
way, the global RequestCollection gets all notifications, so any
request (whether a cluster one or a server one) can subscribe to anything
its interested in.

Signed-off-by: John Spray <john.spray@inktank.com>